requirements.txt
streamlit
matplotlib
streamlit-option-menu
watchdog
gsheetsdb
pyparsing
streamlit-ace
streamlit-discourse
streamlit-disqus
streamlit-elements
streamlit-player
streamlit-quill
streamlit-lottie
seaborn
datetime
folium
keplergl
leafmap
localtileserver
nbserverproxy
owslib
palettable
streamlit-bokeh-events
streamlit-folium
streamlit-keplergl
tropycal
streamlit-echarts
streamlit-extras
streamlit-aggrid
pydeck
streamlit-elements-fluence 
streamlit-on-Hover-tabs
fastapi 
uvicorn 
flask 
asyncpg
python-dotenv
openai
langchain
requests
redis
python-multipart
pandas
llmx
langchain-community
